+++ blockfrost-api/src/Blockfrost/API/Cardano/Accounts.hs
@@ -76,4 +76,23 @@ data AccountsAPI route =
         :> Pagination
         :> Sorting
         :> Get '[JSON] [AccountMir]
+     , _accountAssociatedAddresses
+        :: route
+        :- Summary "Account associated addresses"
+        :> Description "Obtain information about the addresses of a specific account."
+        :> Capture "stake_address" Address
+        :> "addresses"
+        :> Pagination
+        :> Sorting
+        :> Get '[JSON] [AddressAssociated]
+     , _accountAssociatedAssets
+        :: route
+        :- Summary "Assets associated with the account addresses"
+        :> Description "Obtain information about assets associated with addresses of a specific account."
+        :> Capture "stake_address" Address
+        :> "addresses"
+        :> "assets"
+        :> Pagination
+        :> Sorting
+        :> Get '[JSON] [Amount]
     } deriving (Generic)